requirements.txt
akshare
pandas
numpy
SQLAlchemy
psycopg2-binary
python-dateutil
pytz